+++ opentelemetry-aws/src/trace/mod.rs
@@ -1,0 +1,10 @@
+#[cfg(feature = "trace")]
+pub mod id_generator;
+#[cfg(feature = "trace")]
+pub mod xray_propagator;
+
+#[cfg(feature = "trace")]
+pub use xray_propagator::XrayPropagator;
+
+#[cfg(feature = "trace")]
+pub use id_generator::XrayIdGenerator;