+++ opentelemetry-etw-logs/src/logs/mod.rs
@@ -1,0 +1,7 @@
+mod exporter;
+pub use exporter::*;
+
+mod reentrant_logprocessor;
+pub use reentrant_logprocessor::*;
+
+mod converters;