+++ opentelemetry-zpages/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## vNext
 
-WARNING The current version relies on features only in upstream git version. This should be modified before releasing.
+## v0.7.0
+
+### Changed
+
+- Bump opentelemetry version to 0.22, opentelemetry_sdk version to 0.22
 
 ## v0.6.0